Update tiny_mce_gzip.php
I use contao CMS and TinyMCE in a SSL-Proxy environment. It mkaes my server accessable with SSL but sets the SSl-Proxy URL in front of the original server name. In this case TinyMCE will be loaded with wrong path and I can't use TinyMCE in SSL-Proxy environment. With the patch above it works. (see also https://github.com/contao/core/issues/3616#issuecomment-7636126)